[py][BiDi] separate log module from script module and add more tests
