+++ py/test/selenium/webdriver/common/bidi_script_tests.py
@@ -92,3 +92,37 @@ def test_removes_console_message_handler(driver, pages):
 
     WebDriverWait(driver, 5).until(lambda _: len(log_entries2) == 2)
     assert len(log_entries1) == 1
+
+def test_javascript_error_messages(driver, pages):
+    pages.load("bidi/logEntryAdded.html")
+
+    log_entries = []
+    driver.script.add_javascript_error_handler(log_entries.append)
+
+    driver.find_element(By.ID, "jsException").click()
+    WebDriverWait(driver, 5).until(lambda _: log_entries)
+
+    log_entry = log_entries[0]
+    assert log_entry.text == "Error: Not working"
+    assert log_entry.level == "error"
+    assert log_entry.type_ == "javascript"
+
+def test_removes_javascript_message_handler(driver, pages):
+    pages.load("bidi/logEntryAdded.html")
+
+    log_entries1 = []
+    log_entries2 = []
+
+    id = driver.script.add_javascript_error_handler(log_entries1.append)
+    driver.script.add_javascript_error_handler(log_entries2.append)
+
+    driver.find_element(By.ID, "jsException").click()
+    WebDriverWait(driver, 5).until(lambda _: len(log_entries1) and len(log_entries2))
+
+    driver.script.remove_javascript_error_handler(id)
+    driver.find_element(By.ID, "jsException").click()
+
+    WebDriverWait(driver, 5).until(lambda _: len(log_entries2) == 2)
+    assert len(log_entries1) == 1
+
+
